requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
fastapi-sso==0.14.2
loguru==0.7.2
python-dotenv==1.0.1
